+++ jest.config.js
@@ -2,6 +2,10 @@ const esModules = ['escape-string-regexp'].join('|')
 
 module.exports = {
   collectCoverage: true,
+  coveragePathIgnorePatterns: [
+    '/node_modules/',
+    '<rootDir>/lib/helpers/tests.ts',
+  ],
   moduleNameMapper: {
     '\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$':
       '<rootDir>/lib/__mocks__/fileMock.js',
@@ -10,6 +14,6 @@ module.exports = {
   resetMocks: false,
   setupFilesAfterEnv: ['<rootDir>/jest.setup.js'],
   testEnvironment: 'jsdom',
-  testRegex: '(/__tests__/.*)\\.[jt]sx?$',
+  testRegex: '(/__tests__/.*)\\.test\\.[jt]sx?$',
   transformIgnorePatterns: [`node_modules/(?!(${esModules}))`],
 }